requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
seaborn>=0.12
matplotlib>=3.6
mlflow>=2.2
fastapi>=0.110
uvicorn[standard]>=0.22
joblib>=1.2
pytest>=7.0
flake8>=6.0
prometheus-client>=0.16
pydantic>=2.0
requests>=2.31
python-multipart==0.0.21
pillow>=9.0
torch>=1.9.0
torchvision>=0.10.0